Add dummy board node for simulation
The dummy board keeps track of the current state of the board, and responds to
movement messages and requests for scans as appropriate. The idea is that the
movement messages will be produced by another node (probably the one operating
the actuator) once it completes a move, which it will be commanded to do after
the move is determined to be valid.